requirements.txt
grpcio==1.27.1
tensorflow
opencv-python>=4.2.0
SQLAlchemy==1.3.13
torch 
torchvision
tornado
tqdm
Pillow==2.2.2
spicy
imageio
tensorflow-addons
